Added HTML framework for Calculator Buttons
This framework for the Calculator was based on the casio micro mini.
I simply added all of the buttons on the calculator as buttons in HTML.
I also added a screen element using a span. Not sure if this will stay.
Without already knowing how the scripting logic will work, I think this
layout is what I will end up sticking with. Next step is to work on the
CSS and get things looking a lot more like a calculator and less like
a random mess of buttons!