Updated Layout and initialized grid structure
I predictably changed some of the HTML layout to better work with my
CSS choices and also in anticipation of wanting to refer to certain
elements later on in the scripting portion of the project. The basic
structure now works as a grid with the first row being dedicated to the
screen, and subsequent rows being for buttons. The classes I added to
the buttons will be used to structure the grid, while the ids I added
can be used for both individual styling and also as targets for scripts
later on in the project. I still have a lot of CSS work to do for now.
First I am going to set up the grid areas based on the classes so that
each row is set up correctly. After that I will see about getting some
more appropriate spacing for the buttons, and adding some styling such
that they match closer to the target of a casio mico mini.