Finished layout, Removed clutter
Got the calculator's layout to be where I more or less want it to be
before going into the scripting phase. I looked at some old micro minis
and it turns out that many of them had this really interesting gold
flake texture. I am not comfortable enough using images as textures in
css yet to give this a go, but I think I got the essence of the color
accross. I also went ahead and got the screen shape dialed in and added
the number of characters, given my chosen font size, that will be the
maximum for input/output. I may need to revisit this for results that
will be larger by looking into having the font automatically reduce in
size if the output is larger than a certain number of characters. For
now, I want to focus on getting the buttons looking right, mimicking the
oval shape and spacing of the real micro mini.