Finished Styling, Checking Mobile
Used the gap property to space the buttons in a way more similar to the
reference calculator. I also fixed the footer to the bottom of the page,
and made some small font changes. I am mainly committing this now so
that I can see how things look on mobile with the added margin and the
fixed footer. I really want to make sure that these don't interfere with
the mobile user experience since this was a problem for me on the
Etch-A-Sketch project.