Added Media Queries for Mobile Viewing
Predictably the defined height and width broke horribly when being
viewd on mobile so I added a media query that should work better with
mobile screens. For tablets it seems to be the case that the default
size works just fine so I will be keeping that for now. I may do some
further testing but since these are being tested on an Iphone SE which
has a relatively small screen for mobile users I am hoping this will
work well with minimal tweaking required.