Finished Basic Calculator Functions
Knowing that the scripting logic will undoubtedly become more complex,
I have written the basic add subtract divide and multiply functions that
in theory I will be able to use later on. My suspicion is that I will
have to make the numbers pressed be added to an array with an event
listener, then store the operation, then add the second numbers to the
array, then once the user hits the equals sign, have the arrays joined
in such a way that the return value is the number they are looking for.
It sounds challenging but doable.