Added Variables, Event Listeners, Display Function
I added a large number of variables to pass into the different functions
I need for the calculator to work. I added event listeners to the number
buttons, and wrote a display function that takes the text content of the
button that was pressed, puts it into the array for the first set of
numbers, and then displays a joined version of that array on the screen
element by changing it's text content. I think I can see how, from here,
to proceed in regards to many pieces of this project. For example,
I could see how a backspace function that removes the last element from
the array could work, as well as an all clear that empties both arrays
entirely. The main struggle I think will be finding a way to switch
between the num1 array and the num2 array. We shall see.