Wrote functions for displaying output, Fixed CSS
Wrote a series of functions to make the calculator work on a basic
level. For now the way it works is that the textContent of the buttons
that the user clicks on are stored in an array. This array is joined
and used to update the textContent of the screen. When the user clicks
on an operator button, the array is joined once again, the value of
num1 is updated, and we switch to a second array. In addition, the value
of the operator that was selected is stored in an operator variable.
When the user clicks on the equals sign, we take the value of the joined
arrays as num1 and num2 respectively, as well as the value of the
operator stored when they pressed equals, and pass that into the
operate() function, which, depending on the values it receives, calls
the function that corresponds to the appropriate operator (add function
when the value of operator is add). I also updated the CSS to avoid a
small visual issue where the height of the screen would suddenly change
when an operator was pressed because the screen was empty. By setting a
static screen for the height, I should bypass this issue. Next steps are
to implement the backspace, all clear, and decimal functions.