Fixed a few small display bugs
There were a couple of issues that cropped up that needed dealing with
before moving into the next part of the project. First I changed the
content of the screen before any operations take place to a simple
greeting. I might change this later, but I think it works better than
having a bunch of numbers in the screen at first which may be confusing.
I also adjusted the divide function to properly handle an attempt to
divide by zero. I think at this point in the project I should focus on
getting the backspace and all clear buttons to work before moving on to
some more complicated problems.