Added logic for Clear and All Clear buttons
I wrote two functions to handle presses of Clear(C) and All Clear (AC)
buttons. Clear checks to see which array is active, then removes the
last value (most recent) in the array. If the array only has a length of
1, then instead it sets the value to 0. AC on the other hand just
sets the variables num1 and num2 to zero, emptys both arrays, sets the
active array to the first array, and sets the screen content to be
empty. Next step is to change the calculation logic such that the result
can be used as the first value for the next calculation.