Added Decimal and Operator Press Logic
With the default logic, subsequent presses of decimals while decimals
where present allowed the user to input values that couldn't be handled.
Now when the user adds a decimal, it will prevent further decimal
presses until after the post calculation logic is run. Unfortunately for
my sanity and continued existence on this Earth, fixing this behavior
has now added a series of bugs. For one, if you backspace a decimal,
you cannot input another one. Additionally, if trying to utilize the yet
discussed Operator Press Logic, which also has me questioning my
continued membership of the Human club, you will not be able to press a
decimal. I also updated the Operator Press Logic by adding a whole bunch
of complicated nonsense. Basically the way it works is tracking whether
or not you have already pressed an operator button. If you haven't, then
it will allow for default behavior. If however, you have already pressed
an operator, then the operator button tells you to go drunk because you
are home. Unfortunately, this also means that, if you enter a number and
press the operator button one too many times, then the calculator will
not behave correctly at all, causing the person who is trying to create
it to slowly lose his mind as he attempts to build a really something
that someone apparently much smarter than him figured out 65 years ago
before the abomination known as Javascript even existed! And because he
decided to do something just a little different in his code, now he has
essentially wasted several days just coding it this way because the
brightest and best people at The Odin Project decided that this
calculator has to behave specifically in the way that they want it to.
Personally, I think that if you are dumb enough to try to use a
calculator in a way where, before knowing the result of a calculation,
you attempt to start another calculation, then you should go home and
rethink your life. I'm certainly starting to rethink mine!